removed extra parameters, added split to delete-button ids, added new pair to json data
