+++ package.json
@@ -13,22 +13,22 @@
     "url": "git+https://github.com/nss-evening-cohort-10/nutshell-nautilus-explorer/.git"
   },
   "keywords": [],
-  "authors": "Charity Bunyon, Alesha Reed, John Key, Evan Grabenstein",
+  "authors": "Charity Bunyon, Alesha Reed, John Key, Evan Grabenstein, Ivan Phelps, Jamie Phillips, Denise Baker",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/nss-evening-cohort-10/nutshell-nautilus-explorer/issues"
   },
   "homepage": "https://github.com/nss-evening-cohort-10/nutshell-nautilus-explorer#readme",
   "devDependencies": {
-    "@babel/core": "^7.7.2",
-    "@babel/preset-env": "^7.7.1",
+    "@babel/core": "^7.7.4",
+    "@babel/preset-env": "^7.7.4",
     "babel-loader": "^8.0.6",
     "css-loader": "^3.2.0",
-    "eslint": "^6.6.0",
+    "eslint": "^6.7.0",
     "eslint-config-airbnb-base": "^14.0.0",
     "eslint-loader": "^3.0.2",
     "eslint-plugin-import": "^2.18.2",
-    "file-loader": "^4.2.0",
+    "file-loader": "^4.3.0",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
     "mini-css-extract-plugin": "^0.8.0",
@@ -41,9 +41,10 @@
   "dependencies": {
     "axios": "^0.19.0",
     "bootstrap": "^4.3.1",
-    "firebase": "^7.2.2",
+    "firebase": "^7.5.0",
     "jquery": "^3.4.1",
     "popper": "^1.0.1",
     "popper.js": "^1.16.0"
-  }
+  },
+  "author": ""
 }
