+++ src/index.html
@@ -23,7 +23,7 @@
             <a class="nav-link crew" href="#">Crew</a>
           </li>
           <li class="nav-item">
-            <a class="nav-link species" href="#">Species</a>
+            <a class="nav-link view-species" href="#">Species</a>
           </li>
           <li class="nav-item">
             <a class="nav-link destinations" href="#">Destinations</a>
